pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doublef"
version = "0.1.0"
description = "DoubleF: A fast and flexible phase association and earthquake location method"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Longfei Duan", email = "duanlongfei20@mails.ucas.ac.cn" }
]
dependencies = [
    "matplotlib==3.9.4",
    "numpy==1.26.3",
    "obspy==1.4.2",
    "pandas==2.3.0",
    "pyrocko==2025.1.21",
    "scipy==1.13.1",
    "sobol-seq==0.2.0",
    "torch==2.3.1",
    "torchaudio==2.3.1",
    "torchvision==0.18.1",
    "tqdm==4.67.1",
    "natsort==8.4.0",
    "pillow==10.4.0",
]

[project.scripts]
doublef = "doublef.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
